#10 implement PR feedback
