+++ data/migration_controls_template.csv
@@ -1,4 +1,4 @@
-year,in,out
+year,ins,outs
 2020,100000,100000
 2021,100000,100000
 2022,100000,100000
